Remove caching of Poetry install
It is unclear why the CI build fails to find `poetry` even after the
step that runs `pipx ensurepath`, particularly because locally running
the workflow via `act` succeeds. Removing the use of a cache for the
poetry installation itself, so that poetry is always freshly installed,
and thus should always be found on the path.